refactor(allocations): mv usePathHelpers to shared
We want to use this in Projects, too
